refactor: extract CwmpodcastController's HTTP streaming logic into CwmpodcastTrackHelper
~430 lines of HTTP range/streaming transport (serveMedia, resolveLocalPath,
streamLocalFile, streamRemoteFile, resolveSafeRemoteIp, resolveRange,
isLocalHost, terminate, flushBuffers) moved from the controller into
CwmpodcastTrackHelper, which already holds this feature's other
concerns. Pure extraction, no behavior change: the SSRF guard
(resolveSafeRemoteIp), the curl request options (including
CURLOPT_RESOLVE pinning and CURLOPT_FOLLOWLOCATION => false), and the
range-parsing logic are byte-identical before/after the move (verified
by diff), modulo `private function` -> `private static function` and
$this-> -> self::. The controller now gathers the HTTP request details
(host/range/if-modified-since/if-none-match/HEAD) and hands them to
CwmpodcastTrackHelper::serveMedia(), the one new public entry point;
every other moved method stays private to the helper, unreachable from
outside it. serveMedia()'s dispatcher return; is deliberately excluded
from the terminate-discipline test — a hand-off to a callee that always
terminates, not a terminal branch.

All 17 existing regression tests (resolveRange edge cases, isLocalHost
port handling, the terminate-discipline source guard, the curl-
availability guard, and the SSRF host-safety cases) pass unmodified
apart from their class references, confirming no invariant broke in
the move.

Live-tested on j5-dev against a local (non-YouTube) media file's
tracking URL: plain GET returned 200 with Accept-Ranges: bytes and the
correct Content-Length; a Range: bytes=0-1023 request returned 206 with
Content-Range: bytes 0-1023/46247281 and a 1024-byte body; HEAD
returned 200 with the same headers. The remote-proxy path (streamRemoteFile)
needs a real external host and wasn't live-tested.

Fixes #1446
